move Header component into its own folder
